pibridge-serdev: change datatype for data length in pibridge_send()
Allow to send more than 255 byte at once with pibridge_send(). This is
needed for larger frames e.g. when updating the firmware.

+++ drivers/tty/serdev/pibridge.c
@@ -151,7 +151,7 @@ static void pibridge_remove(struct serdev_device *serdev)
 
 /*****************/
 
-int pibridge_send(void *buf, u8 len)
+int pibridge_send(void *buf, u32 len)
 {
 	struct pibridge *pi = pibridge_s;
 	struct serdev_device *serdev = pi->serdev;
